requirements.txt
streamlit
pdfplumber
python-docx
sentence-transformers
faiss-cpu
torch
transformers
numpy

